requirements.txt
streamlit
PyPDF2
python-dotenv
openai
moviepy
Pillow
imageio
imageio-ffmpeg
gTTS==2.3.2
pydub==0.25.1
opencv-python-headless>=4.8.0
gdown>=4.7.1
requests>=2.31.0 